backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
chromadb==0.4.15
google-generativeai==0.3.0
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.4
PyJWT==1.7.1
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
gunicorn==21.2.0
